Add user authentication to backend
The system is able to enforce three roles: reader, writer, and owner; it
throw errors when a user tries to perform an action it is not authorized to.

It checks role based on a passed in authentication token (which is obtained on
the client side by communicating with Firestore servers and having a Google OAuth
token).

Also refactors some functions to be object-based rather than global-based.